Refactor app lifecycle hooks into events that can be subscribed to externally
We're moving away from having call site-specific logic inside the composer package, to emitting events that will be consumed by call site interfaces on an as-needed basis
