+++ packages/composer/action-creators/AppInitActionCreators.js
@@ -2,7 +2,7 @@ import cloneDeep from 'lodash.clonedeep';
 import twitterText from 'twitter-text';
 import AppDispatcher from '../dispatcher';
 import { ActionTypes, AttachmentTypes, NotificationScopes } from '../AppConstants';
-import AppHooks from '../utils/lifecycle-hooks';
+import AppHooks from '../utils/LifecycleHooks';
 import AppActionCreators from '../action-creators/AppActionCreators';
 import NotificationActionCreators from '../action-creators/NotificationActionCreators';
 import { getFileTypeFromPath } from '../utils/StringUtils';
