+++ packages/composer/index.js
@@ -1,2 +1,3 @@
 export default from './components/App';
 export { formatInputData } from './utils/DataImportUtils';
+export { default as events } from './utils/Events';
