+++ packages/composer/utils/Events.js
@@ -1,0 +1,3 @@
+import Events from 'event-pubsub';
+
+export default new Events();
